Try to fix build.

+++ src/lib/encoder.cc
@@ -154,7 +154,7 @@ Encoder::process_end ()
 		/* Put audio in where there is none at all */
 		int64_t af = video_frames_to_audio_frames (_video_frames_out, 48000, _film->dcp_frame_rate ());
 		while (af) {
-			int64_t const this_time = min (af, 24000L);
+			int64_t const this_time = min (af, static_cast<int64_t> (24000));
 			shared_ptr<AudioBuffers> out (new AudioBuffers (_film->minimum_audio_channels(), this_time));
 			out->make_silent ();
 			out->set_frames (this_time);
